Implement classes allow to store and access some data safely.
In order to work safely with listeners, especially nested (e.g. the
build listener should set up the task execution graph listener), we
must have an api that allows us to keep track of consistency of data
we are collecting. And as far as we are not allowed to throw exceptions
from listeners we must have another way to detect whether something
went wrong on some nested level.
Implemented api allows us to easily combine listeners by storing data
as properties that can be either correct value or failure that we can
rethrow to an upper level.