Implement concrete properties related to build listening.
Consider InfoProvider as a property that should be
directly set from the outside. In this case, Failure is a
String showing whether this property was set before the get
method is called.

Consider InfoCollector as a property that contains
other properties and collects them in a resulting object on
the get method is called. In this case, Failure is a List
of Strings that contains nested properties failures if any