Add extensions to gradle Tasks allowing to add listeners
Gradle does not provide Task#addTaskExecutionListener, only
TaskExecutionGraph#addTaskExecutionListener, where corresponding
listener methods are called with the Task instance as a parameter.
It's not so convenient, because in this case we must assign some key
to each Task and store them in some storage for consistency between
beforeExecute() and afterExecute().
Now we can add task execution listener to the Task directly and the
listener will know about the task he is listening to.