Rework collecting implementation
What were changed:
1. Now Info data classes are collected in BuildCompletionListener,
but not in every listener separately. This makes it easier for us
to add new intermediate collectors because in this case we're not
limited with a data-classes constructors and we can add all new info
we want before build completed

2. All listeners are added in one place. This allows us to fully
separate the logic of info collecting, now the listeners are
independent of each other and follow the Single responsibility principle

3. Consider 1 and 2, we refused to implement individual task listeners

4. In order to implement collecting info about tasks inputs,
tasks info collecting logic was separated on two parts:
collecting common task info (path, name, time measuring) and collecting
task state info.