Implement database management
Implementation notes:
1. H2 is used as underlying database
2. The `squash` library is used as a typed wrapper over the database.
3. We are not allowed to run a transaction inside another transaction
and we have to close first transaction before we start second one.
4. Our BuildInfo data-class contains other nested complex data-classes
and we have to store it in multiple tables.
5. Based on 1 and 2, we can't fetch info about nested field while
constructing outer data-class, so we have to fetch all needed data
from one table, close transaction and then start fetching info about
children. Classes in TableModels.kt is responsible for this
intermediate state.